Separation of concerns - MRUK extended methods, GameAnchorSelection
