+++ Assets/Scripts/MRUK/GameAnchorSelection.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6b8ef1c05e7d2bf4e9e80665b775223a
